+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/Reports.kt
@@ -33,7 +33,10 @@ data class JacocoReport(
 )
 
 enum class ReportType {
-    HTML, XML, CSV
+    HTML,
+    XML,
+    @Deprecated("CSV will be removed soon.")
+    CSV,
 }
 
 data class Violations(